pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "raceocr"
version = "0.0.1"
description = "Race OCR inference CLI (YOLO + PaddleOCR)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "ultralytics>=8.3,<9",
  "pillow>=10,<12",
  "paddleocr>=3,<4",
  "paddlepaddle==3.2.2",
  "requests<3",
  "urllib3<2.3",
]

[project.scripts]
raceocr = "raceocr.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]